% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{as_correlation_matrix}
\alias{as_correlation_matrix}
\title{Convert matrix to CorrelationMatrix object}
\usage{
as_correlation_matrix(matrix, n_hypotheses = 0L, n_analyses = 0L)
}
\arguments{
\item{matrix}{A numeric correlation matrix}

\item{n_hypotheses}{Integer number of hypotheses (optional)}

\item{n_analyses}{Integer number of analyses (optional)}
}
\value{
A CorrelationMatrix S7 object
}
\description{
Converts a regular numeric matrix to a CorrelationMatrix S7 object with validation.
}
\examples{
corr_mat <- matrix(c(1, 0.3, 0.3, 1), nrow = 2)
corr_obj <- as_correlation_matrix(corr_mat, n_hypotheses = 1, n_analyses = 2)
print(corr_obj)

}
