% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_corr.R
\name{d_corr}
\alias{d_corr}
\title{Convert event matrix to correlation matrix}
\usage{
d_corr(D)
}
\arguments{
\item{D}{Event matrix.}
}
\value{
Correlation matrix.
}
\description{
Convert event matrix to correlation matrix
}
\examples{
wpgsd:::d_corr(D = matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), nrow = 3))
}
