% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bounds.R
\name{generate_bounds}
\alias{generate_bounds}
\title{Compute p value boundaries of the parametric MTP method with overall alpha spending for all hypotheses}
\usage{
generate_bounds(
  type = 1,
  k = 2,
  w = w,
  m = m,
  corr = corr,
  alpha = 0.025,
  cum_alpha = NULL,
  maxpts = 50000,
  abseps = 1e-05,
  tol = 1e-10,
  sf = gsDesign::sfHSD,
  sfparm = -4,
  t = c(0.5, 1),
  ...
)
}
\arguments{
\item{type}{Boundary type.
0 = Bonferroni. Separate alpha spending for each hypotheses.
1 = Fixed alpha spending for all hypotheses. Method 3a in the manuscript.
2 = Overall alpha spending for all hypotheses. Method 3b in the manuscript.
3 = Separate alpha spending for each hypotheses. Method 3c in the manuscript.}

\item{k}{Number of analyses up to the current analysis.}

\item{w}{Initial weights}

\item{m}{Transition matrix}

\item{corr}{Correlation matrix of all test statistics up to the current analysis. dim = k*length(w)}

\item{alpha}{Overall alpha}

\item{cum_alpha}{Cumulative alpha spent at each analysis. Only required for type=1.}

\item{maxpts}{GenzBretz function maximum number of function values as integer}

\item{abseps}{GenzBretz function absolute error tolerance}

\item{tol}{find root tolerance}

\item{sf}{A list of alpha spending functions to spend alpha for each hypotheses.
If type = 0 or 3 then length equals to number of hypotheses.
If type = 1 then sf is not needed.
If type = 2 then only the first component is used.}

\item{sfparm}{A list of parameters to be supplied to sfs.
If type = 0 or 3 then length equals to number of hypotheses.
If type = 1 then sfparm is not needed.
If type = 2 then only the first component is used.}

\item{t}{A list of information fraction used for alpha spending,
may be different from the actual information fraction.
Each component corresponds to a hypothesis.
If type = 0 or 3 then length equals to number of hypotheses.
If type = 1 then t is not needed.
If type = 2 then only the first component is used.}

\item{...}{additional arguments}
}
\value{
A \code{tibble} with \code{k*(2^(n_hypotheses-1))} rows of p-value boundaries.
Inflation factor is also provided if type=3.
}
\description{
Compute p value boundaries of the parametric MTP method with overall alpha spending for all hypotheses
}
\examples{
library(tibble)

# build the transition matrix
m <- matrix(c(0, 0.5, 0.5, 
              0.5, 0, 0.5,
              0.5,0.5,0), nrow = 3, byrow = TRUE)
# initialize weights
w <- c(1/3, 1/3, 1/3)

# input information fraction
IF_IA <- c(155/305, 160/320, 165/335) 

# input event count of intersection of paired hypotheses - Table 2
event <- tribble(~H1, ~H2, ~Analysis, ~Event,
                 1, 1, 1, 155,
                 2, 2, 1, 160,
                 3, 3, 1, 165,
                 1, 2, 1, 85,
                 1, 3, 1, 85,
                 2, 3, 1, 85,
                 1, 1, 2, 305,
                 2, 2, 2, 320,
                 3, 3, 2, 335,
                 1, 2, 2, 170,
                 1, 3, 2, 170,
                 2, 3, 2, 170) 

# generate correlation from events
gs_corr <- generate_corr(event)

# generate bounds
generate_bounds(type = 3,
                k = 2, 
                w = w, 
                m = m, 
                corr = gs_corr, 
                alpha = 0.025, 
                sf = list(gsDesign::sfLDOF, gsDesign::sfLDOF, gsDesign::sfLDOF), 
                sfparm = list(0, 0, 0), 
                t = list(c(IF_IA[1], 1), c(IF_IA[2], 1), c(IF_IA[3], 1))) 
                
}
