% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_event_table.R
\name{generate_event_table}
\alias{generate_event_table}
\title{Generate table of event counts from ADSL and ADTTE datasets}
\usage{
generate_event_table(paths, h_select, adsl_name, adtte_name, key_var, cnsr_var)
}
\arguments{
\item{paths}{A vector of paths for analysis datasets.
Length should be equal to the number of analyses completed.}

\item{h_select}{Selection criterion for each hypothesis.
Should be a tibble containing 2 columns: \code{Hypothesis} and \code{Crit}.}

\item{adsl_name}{SAS dataset name for subject-level analysis data.
Usually it is \code{"adsl"}.}

\item{adtte_name}{SAS dataset name for time-to-event analysis data.
Usually it is \code{"adtte"}.}

\item{key_var}{Key variable to join the \code{adsl} and \code{adtte} datasets.
For example, \code{"USUBJID"} or \code{"SUBJID"}.}

\item{cnsr_var}{Variable to indicate censoring (\code{1} = censor; \code{0} = event).
For example, \code{"CNSR"}.}
}
\value{
A list with two components:
\itemize{
\item \code{event}: an event count table as input for \code{\link[=generate_bounds]{generate_bounds()}}.
\item \code{dsets}: analysis datasets of each hypothesis.
}
}
\description{
Generate table of event counts from ADSL and ADTTE datasets
}
\examples{
library(dplyr)

paths <- system.file("extdata/", package = "wpgsd")

# Selection criteria for each hypothesis
h_select <- tibble::tribble(
  ~Hypothesis, ~Crit,
  1, "PARAMCD == 'OS' & TRT01P \%in\% c('Xanomeline High Dose', 'Placebo')",
  2, "PARAMCD == 'OS' & TRT01P \%in\% c('Xanomeline Low Dose', 'Placebo')"
)

event <- generate_event_table(paths, h_select,
  adsl_name = "adsl", adtte_name = "adtte",
  key_var = "USUBJID", cnsr_var = "CNSR"
)$event

event \%>\%
  gt::gt() \%>\%
  gt::tab_header(title = "Event Count - Computed from SAS Datasets Example")
}
