% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_event_cc.R
\name{generate_event_table_cc}
\alias{generate_event_table_cc}
\title{This function generates a table of events for given experimental arms and a control group based on specified hypotheses.}
\usage{
generate_event_table_cc(event, hypothesis)
}
\arguments{
\item{event}{A dataframe containing the following columns:
\itemize{
\item \code{Population}: A character vector listing the population groups (e.g., experimental arms and control).
\item \code{IA}: A numeric vector indicating the number of events observed in each group during interim analysis.
\item \code{FA}: A numeric vector indicating the number of events observed in each group during final analysis.
The dataframe must contain at least these columns and can include additional analysis columns as needed.
}}

\item{hypothesis}{A list containing hypotheses specifying comparisons between experimental arms and the control group,
as well as comparisons among experimental arms.}
}
\value{
A dataframe with columns:
\itemize{
\item \code{one_hypothesis}: The index of the first selected hypothesis from the provided list.
\item \code{another_hypothesis}: The index of the second selected hypothesis from the provided list.
\item \code{analysis}: The index indicating which analysis is being performed (e.g., interim or final).
\item \code{common_events}: The calculated number of common events associated with the selected hypotheses.
}
}
\description{
This function generates a table of events for given experimental arms and a control group based on specified hypotheses.
}
\examples{
#------------------------Example of IA and FA
event <- data.frame(
  Population = c("Experimental 1", "Experimental 2", "Experimental 3", "Control"),
  IA = c(70, 75, 80, 85), # Interim Analysis values indicating the number of events observed in each group
  FA = c(135, 150, 165, 170)
)

hypothesis <- list(
  H1 = "Experimental 1 vs. Control",
  H2 = "Experimental 2 vs. Control",
  H3 = "Experimental 1 vs. Experimental 2"
)

generate_event_table_cc(event, hypothesis)

#----------------------Example of two IAs and FA
event <- data.frame(
  Population = c("Experimental 1", "Experimental 2", "Experimental 3", "Control"),
  IA1 = c(70, 75, 80, 85), # First Interim Analysis values indicating the number of events observed in each group
  IA2 = c(90, 95, 100, 105), # Second Interim Analysis values indicating the number of events observed in each group
  FA = c(135, 150, 165, 170)
)

hypothesis <- list(
  H1 = "Experimental 1 vs. Control",
  H2 = "Experimental 2 vs. Control",
  H3 = "Experimental 1 vs. Experimental 2"
)

generate_event_table_cc(event, hypothesis)
}
