% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_event_ol.R
\name{generate_event_table_ol}
\alias{generate_event_table_ol}
\title{This function generates a table of events for specified populations based on the provided hypotheses.}
\usage{
generate_event_table_ol(event, hypothesis)
}
\arguments{
\item{event}{dataframe should have the following structure:
\itemize{
\item \code{Population}: A character vector indicating the population groups (e.g., "Population 1", "Population 2", "Population 1 Intersection 2", and "Overall population").
\item \code{IA}: Numeric vector indicating the number of events observed in each group during interim analysis.
\item \code{FA}: Numeric vector indicating the number of events observed in each group during final analysis.
The dataframe must contain at least these columns and can include additional analysis columns as needed.
}}

\item{hypothesis}{A list of strings where each item represents a hypothesis regarding efficacy, formatted as follows:
- H1: "Efficacy in Population 1"
- H2: "Efficacy in Population 2"
- H3: "Efficacy in Overall population"
Each hypothesis is used for comparisons in the generated event table.}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{one_hypothesis}: The index of the first selected hypothesis from the provided list.
\item \code{another_hypothesis}: The index of the second selected hypothesis from the provided list.
\item \code{analysis}: The index indicating which analysis is being performed (e.g., interim or final).
\item \code{common_events}: The calculated number of common events associated with the selected hypotheses.
}
}
\description{
This function generates a table of events for specified populations based on the provided hypotheses.
}
\examples{
#------------------------Example of IA and FA
event <- data.frame(
  Population = c("Population 1", "Population 2", "Population 1 Intersection 2", "Overall population"),
  IA = c(100, 110, 80, 225), # Interim Analysis values indicating the number of events observed in each group
  FA = c(200, 220, 160, 450)
)

hypothesis <- list(
  H1 = "Efficacy in Population 1",
  H2 = "Efficacy in Population 2",
  H3 = "Efficacy in Overall population"
)

generate_event_table_ol(event, hypothesis)

#----------------------Example of two IAs and FA
event <- data.frame(
  Population = c("Population 1", "Population 2", "Population 1 Intersection 2", "Overall population"),
  IA1 = c(100, 110, 80, 225), # First Interim Analysis values indicating the number of events observed in each group
  IA2 = c(120, 130, 90, 240), # Second Interim Analysis values indicating the number of events observed in each group
  FA = c(200, 220, 160, 450)
)

hypothesis <- list(
  H1 = "Efficacy in Population 1",
  H2 = "Efficacy in Population 2",
  H3 = "Efficacy in Overall population"
)

generate_event_table_ol(event, hypothesis)

}
