% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_path.R
\name{generate_path}
\alias{generate_path}
\title{Generate path for Merck studies on CPI drive}
\usage{
generate_path(env, project, deliverables)
}
\arguments{
\item{env}{environment. prod or test.}

\item{project}{project. e.g. mk3475-tnbc}

\item{deliverables}{a vector containing deliverable folder name of each analysis. e.g. c("prot001-ia01", "prot001-ia02")}
}
\value{
data path(s) pointing to the folders containing data sets of each analysis
}
\description{
Generate path for Merck studies on CPI drive
}
\examples{
paths <- generate_path(env = "prod",
                       project = "mk3475-tnbc",
                       deliverables = c("prot355-ia01-eff", "prot355-ia02-eff"))
paths

}
