% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{print.CorrelationMatrix}
\alias{print.CorrelationMatrix}
\title{Print method for CorrelationMatrix}
\arguments{
\item{x}{A CorrelationMatrix S7 object}

\item{...}{Additional arguments passed to print methods}
}
\value{
Invisibly returns the input object
}
\description{
Print method for CorrelationMatrix
}
