% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgsd-package.R
\docType{package}
\name{wpgsd-package}
\alias{wpgsd}
\alias{wpgsd-package}
\title{wpgsd: Weighted Parametric Group Sequential Design}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Adjusted inference for weighted parametric group sequential design. Weighted parametric group sequential design (WPGSD) by Anderson et al. (2022) \doi{10.1002/bimj.202100085} allows one to take advantage of the known correlation structure in constructing efficacy bounds to control family-wise error rate (FWER) for a group sequential design. Here, correlation may be due to common observations in nested populations, due to common observations in overlapping populations, or due to common observations in the control arm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/wpgsd/}
  \item \url{https://github.com/Merck/wpgsd}
  \item Report bugs at \url{https://github.com/Merck/wpgsd/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Zifang Guo \email{zifang.guo@merck.com}
  \item Jing Zhao \email{jing_zhaox@merck.com}
  \item Linda Sun \email{linda_sun@merck.com}
  \item Yi Cui \email{yi.cui@merck.com}
  \item Larry Leon \email{larry.leon2@merck.com}
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
